feat: refactor context
